+++ pystac/extensions/label.py
@@ -639,13 +639,21 @@ class LabelExtension(ExtensionManagementMixin[pystac.Item]):
         return SCHEMA_URI
 
     @classmethod
-    def ext(cls, obj: pystac.Item) -> "LabelExtension":
+    def ext(cls, obj: pystac.Item, add_if_missing: bool = False) -> "LabelExtension":
         """Extends the given STAC Object with properties from the :stac-ext:`Label
         Extension <label>`.
 
         This extension can be applied to instances of :class:`~pystac.Item`.
         """
-        return cls(obj)
+        if isinstance(obj, pystac.Item):
+            if add_if_missing:
+                cls.add_to(obj)
+            cls.validate_has_extension(obj)
+            return cls(obj)
+        else:
+            raise pystac.ExtensionTypeError(
+                f"Label extension does not apply to type {type(obj)}"
+            )
 
 
 class LabelExtensionHooks(ExtensionHooks):