+++ tests/extensions/test_sat.py
@@ -7,6 +7,7 @@ import unittest
 import pystac
 from pystac.extensions import sat
 from pystac.extensions.sat import SatExtension
+from tests.utils import TestCases
 
 
 def make_item() -> pystac.Item:
@@ -25,6 +26,7 @@ class SatTest(unittest.TestCase):
     def setUp(self) -> None:
         super().setUp()
         self.item = make_item()
+        self.sentinel_example_uri = TestCases.get_path("data-files/sat/sentinel-1.json")
 
     def test_stac_extensions(self) -> None:
         self.assertTrue(SatExtension.has_extension(self.item))
@@ -90,7 +92,7 @@ class SatTest(unittest.TestCase):
             "geometry": None,
             "links": [],
             "assets": {},
-            "stac_extensions": ["sat"],
+            "stac_extensions": [SatExtension.get_schema_uri()],
         }
         item = pystac.Item.from_dict(d)
         self.assertEqual(orbit_state, SatExtension.ext(item).orbit_state)
@@ -121,3 +123,40 @@ class SatTest(unittest.TestCase):
         SatExtension.ext(self.item).relative_orbit = None
         self.assertIsNone(SatExtension.ext(self.item).relative_orbit)
         self.item.validate()
+
+    def test_extension_not_implemented(self) -> None:
+        # Should raise exception if Item does not include extension URI
+        item = pystac.Item.from_file(self.sentinel_example_uri)
+        item.stac_extensions.remove(SatExtension.get_schema_uri())
+
+        with self.assertRaises(pystac.ExtensionNotImplemented):
+            _ = SatExtension.ext(item)
+
+        # Should raise exception if owning Item does not include extension URI
+        asset = item.assets["measurement_iw1_vh"]
+
+        with self.assertRaises(pystac.ExtensionNotImplemented):
+            _ = SatExtension.ext(asset)
+
+        # Should succeed if Asset has no owner
+        ownerless_asset = pystac.Asset.from_dict(asset.to_dict())
+        _ = SatExtension.ext(ownerless_asset)
+
+    def test_item_ext_add_to(self) -> None:
+        item = pystac.Item.from_file(self.sentinel_example_uri)
+        item.stac_extensions.remove(SatExtension.get_schema_uri())
+        self.assertNotIn(SatExtension.get_schema_uri(), item.stac_extensions)
+
+        _ = SatExtension.ext(item, add_if_missing=True)
+
+        self.assertIn(SatExtension.get_schema_uri(), item.stac_extensions)
+
+    def test_asset_ext_add_to(self) -> None:
+        item = pystac.Item.from_file(self.sentinel_example_uri)
+        item.stac_extensions.remove(SatExtension.get_schema_uri())
+        self.assertNotIn(SatExtension.get_schema_uri(), item.stac_extensions)
+        asset = item.assets["measurement_iw1_vh"]
+
+        _ = SatExtension.ext(asset, add_if_missing=True)
+
+        self.assertIn(SatExtension.get_schema_uri(), item.stac_extensions)